+++ packages/qdadm/src/components/layout/defaults/index.js
@@ -11,7 +11,6 @@ export { default as DefaultMenu } from './DefaultMenu.vue'
 export { default as DefaultFooter } from './DefaultFooter.vue'
 export { default as DefaultUserInfo } from './DefaultUserInfo.vue'
 export { default as DefaultBreadcrumb } from './DefaultBreadcrumb.vue'
-export { default as DefaultToaster } from './DefaultToaster.vue'
 
 // FormLayout defaults
 export { default as DefaultFormActions } from './DefaultFormActions.vue'